+++ app.py
@@ -1,4 +1,4 @@
-from src.main.classes.Zillow import search 
+from src.main.classes.zillow.Zillow import search 
 
 if __name__ == '__main__':
     result = search("X1-ZWz17jvxcowwln_7lwvq", "60611", "420 E Ohio St")
